[WIP] move ACLK_Query to separate files
