+++ aclk/aclk_common.c
@@ -2,6 +2,14 @@
 
 #include "../daemon/common.h"
 
+netdata_mutex_t aclk_shared_state_mutex = NETDATA_MUTEX_INITIALIZER;
+
+struct aclk_shared_state aclk_shared_state = {
+    .metadata_submitted = ACLK_METADATA_REQUIRED,
+    .agent_state = AGENT_INITIALIZING,
+    .last_popcorn_interrupt = 0
+};
+
 struct {
     ACLK_PROXY_TYPE type;
     const char *url_str;
