+++ aclk/agent_cloud_link.c
@@ -3,6 +3,7 @@
 #include "libnetdata/libnetdata.h"
 #include "agent_cloud_link.h"
 #include "aclk_lws_https_client.h"
+#include "aclk_query.h"
 #include "aclk_common.h"
 #include "aclk_stats.h"
 
@@ -11,7 +12,7 @@ int aclk_shutting_down = 0;
 // Other global state
 static int aclk_subscribed = 0;
 static int aclk_disable_single_updates = 0;
-static int waiting_init = 1;
+int waiting_init = 1;
 static char *aclk_username = NULL;
 static char *aclk_password = NULL;
 
@@ -24,28 +25,13 @@ usec_t aclk_session_us = 0;         // Used by the mqtt layer
 time_t aclk_session_sec = 0;        // Used by the mqtt layer
 
 static netdata_mutex_t aclk_mutex = NETDATA_MUTEX_INITIALIZER;
-static netdata_mutex_t aclk_shared_state_mutex = NETDATA_MUTEX_INITIALIZER;
-static netdata_mutex_t queue_mutex = NETDATA_MUTEX_INITIALIZER;
 static netdata_mutex_t collector_mutex = NETDATA_MUTEX_INITIALIZER;
 
 #define ACLK_LOCK netdata_mutex_lock(&aclk_mutex)
 #define ACLK_UNLOCK netdata_mutex_unlock(&aclk_mutex)
 
-#define ACLK_SHARED_STATE_LOCK netdata_mutex_lock(&aclk_shared_state_mutex)
-#define ACLK_SHARED_STATE_UNLOCK netdata_mutex_unlock(&aclk_shared_state_mutex)
-
 #define COLLECTOR_LOCK netdata_mutex_lock(&collector_mutex)
 #define COLLECTOR_UNLOCK netdata_mutex_unlock(&collector_mutex)
-
-#define ACLK_QUEUE_LOCK netdata_mutex_lock(&queue_mutex)
-#define ACLK_QUEUE_UNLOCK netdata_mutex_unlock(&queue_mutex)
-
-pthread_cond_t query_cond_wait = PTHREAD_COND_INITIALIZER;
-pthread_mutex_t query_lock_wait = PTHREAD_MUTEX_INITIALIZER;
-
-#define QUERY_THREAD_LOCK pthread_mutex_lock(&query_lock_wait);
-#define QUERY_THREAD_UNLOCK pthread_mutex_unlock(&query_lock_wait)
-#define QUERY_THREAD_WAKEUP pthread_cond_signal(&query_cond_wait)
 
 void lws_wss_check_queues(size_t *write_len, size_t *write_len_bytes, size_t *read_len);
 void aclk_lws_wss_destroy_context();
@@ -68,48 +54,6 @@ struct _collector {
 };
 
 struct _collector *collector_list = NULL;
-
-struct aclk_query {
-    time_t created;
-    time_t run_after; // Delay run until after this time
-    ACLK_CMD cmd;     // What command is this
-    char *topic;      // Topic to respond to
-    char *data;       // Internal data (NULL if request from the cloud)
-    char *msg_id;     // msg_id generated by the cloud (NULL if internal)
-    char *query;      // The actual query
-    u_char deleted;   // Mark deleted for garbage collect
-    struct aclk_query *next;
-};
-
-struct aclk_query_queue {
-    struct aclk_query *aclk_query_head;
-    struct aclk_query *aclk_query_tail;
-    uint64_t count;
-} aclk_queue = { .aclk_query_head = NULL, .aclk_query_tail = NULL, .count = 0 };
-
-// This is to hold data that can be accessed by any
-// thread and thus protected by mutex
-// instead of reusing mutex for different vars
-// we group them together here
-static struct aclk_shared_state {
-    ACLK_METADATA_STATE metadata_submitted;
-    AGENT_STATE agent_state;
-    time_t last_popcorn_interrupt;
-} aclk_shared_state = {
-    .metadata_submitted = ACLK_METADATA_REQUIRED,
-    .agent_state = AGENT_INITIALIZING,
-    .last_popcorn_interrupt = 0
-};
-
-struct aclk_query_thread {
-    netdata_thread_t *thread;
-    int idx;
-};
-
-struct aclk_query_threads {
-    struct aclk_query_thread **thread_list;
-    int count;
-};
 
 char *create_uuid()
 {
@@ -239,220 +183,6 @@ unsigned long int aclk_reconnect_delay(int mode)
     }
 
     return delay;
-}
-
-/*
- * Free a query structure when done
- */
-
-void aclk_query_free(struct aclk_query *this_query)
-{
-    if (unlikely(!this_query))
-        return;
-
-    freez(this_query->topic);
-    if (likely(this_query->query))
-        freez(this_query->query);
-    if (likely(this_query->data))
-        freez(this_query->data);
-    if (likely(this_query->msg_id))
-        freez(this_query->msg_id);
-    freez(this_query);
-}
-
-// Returns the entry after which we need to create a new entry to run at the specified time
-// If NULL is returned we need to add to HEAD
-// Need to have a QUERY lock before calling this
-
-struct aclk_query *aclk_query_find_position(time_t time_to_run)
-{
-    struct aclk_query *tmp_query, *last_query;
-
-    // Quick check if we will add to the end
-    if (likely(aclk_queue.aclk_query_tail)) {
-        if (aclk_queue.aclk_query_tail->run_after <= time_to_run)
-            return aclk_queue.aclk_query_tail;
-    }
-
-    last_query = NULL;
-    tmp_query = aclk_queue.aclk_query_head;
-
-    while (tmp_query) {
-        if (tmp_query->run_after > time_to_run)
-            return last_query;
-        last_query = tmp_query;
-        tmp_query = tmp_query->next;
-    }
-    return last_query;
-}
-
-// Need to have a QUERY lock before calling this
-struct aclk_query *
-aclk_query_find(char *topic, char *data, char *msg_id, char *query, ACLK_CMD cmd, struct aclk_query **last_query)
-{
-    struct aclk_query *tmp_query, *prev_query;
-    UNUSED(cmd);
-
-    tmp_query = aclk_queue.aclk_query_head;
-    prev_query = NULL;
-    while (tmp_query) {
-        if (likely(!tmp_query->deleted)) {
-            if (strcmp(tmp_query->topic, topic) == 0 && (!query || strcmp(tmp_query->query, query) == 0)) {
-                if ((!data || (data && strcmp(data, tmp_query->data) == 0)) &&
-                    (!msg_id || (msg_id && strcmp(msg_id, tmp_query->msg_id) == 0))) {
-                    if (likely(last_query))
-                        *last_query = prev_query;
-                    return tmp_query;
-                }
-            }
-        }
-        prev_query = tmp_query;
-        tmp_query = tmp_query->next;
-    }
-    return NULL;
-}
-
-/*
- * Add a query to execute, the result will be send to the specified topic
- */
-
-int aclk_queue_query(char *topic, char *data, char *msg_id, char *query, int run_after, int internal, ACLK_CMD aclk_cmd)
-{
-    struct aclk_query *new_query, *tmp_query;
-
-    // Ignore all commands while we wait for the agent to initialize
-    if (unlikely(waiting_init))
-        return 1;
-
-    run_after = now_realtime_sec() + run_after;
-
-    ACLK_QUEUE_LOCK;
-    struct aclk_query *last_query = NULL;
-
-    tmp_query = aclk_query_find(topic, data, msg_id, query, aclk_cmd, &last_query);
-    if (unlikely(tmp_query)) {
-        if (tmp_query->run_after == run_after) {
-            ACLK_QUEUE_UNLOCK;
-            QUERY_THREAD_WAKEUP;
-            return 0;
-        }
-
-        if (last_query)
-            last_query->next = tmp_query->next;
-        else
-            aclk_queue.aclk_query_head = tmp_query->next;
-
-        debug(D_ACLK, "Removing double entry");
-        aclk_query_free(tmp_query);
-        aclk_queue.count--;
-    }
-
-    if (aclk_stats_enabled) {
-        ACLK_STATS_LOCK;
-        aclk_metrics_per_sample.queries_queued++;
-        ACLK_STATS_UNLOCK;
-    }
-
-    new_query = callocz(1, sizeof(struct aclk_query));
-    new_query->cmd = aclk_cmd;
-    if (internal) {
-        new_query->topic = strdupz(topic);
-        if (likely(query))
-            new_query->query = strdupz(query);
-    } else {
-        new_query->topic = topic;
-        new_query->query = query;
-        new_query->msg_id = msg_id;
-    }
-
-    if (data)
-        new_query->data = strdupz(data);
-
-    new_query->next = NULL;
-    new_query->created = now_realtime_sec();
-    new_query->run_after = run_after;
-
-    debug(D_ACLK, "Added query (%s) (%s)", topic, query ? query : "");
-
-    tmp_query = aclk_query_find_position(run_after);
-
-    if (tmp_query) {
-        new_query->next = tmp_query->next;
-        tmp_query->next = new_query;
-        if (tmp_query == aclk_queue.aclk_query_tail)
-            aclk_queue.aclk_query_tail = new_query;
-        aclk_queue.count++;
-        ACLK_QUEUE_UNLOCK;
-        QUERY_THREAD_WAKEUP;
-        return 0;
-    }
-
-    new_query->next = aclk_queue.aclk_query_head;
-    aclk_queue.aclk_query_head = new_query;
-    aclk_queue.count++;
-
-    ACLK_QUEUE_UNLOCK;
-    QUERY_THREAD_WAKEUP;
-    return 0;
-}
-
-/*
- * Get the next query to process - NULL if nothing there
- * The caller needs to free memory by calling aclk_query_free()
- *
- *      topic
- *      query
- *      The structure itself
- *
- */
-struct aclk_query *aclk_queue_pop()
-{
-    struct aclk_query *this_query;
-
-    ACLK_QUEUE_LOCK;
-
-    if (likely(!aclk_queue.aclk_query_head)) {
-        ACLK_QUEUE_UNLOCK;
-        return NULL;
-    }
-
-    this_query = aclk_queue.aclk_query_head;
-
-    // Get rid of the deleted entries
-    while (this_query && this_query->deleted) {
-        aclk_queue.count--;
-
-        aclk_queue.aclk_query_head = aclk_queue.aclk_query_head->next;
-
-        if (likely(!aclk_queue.aclk_query_head)) {
-            aclk_queue.aclk_query_tail = NULL;
-        }
-
-        aclk_query_free(this_query);
-
-        this_query = aclk_queue.aclk_query_head;
-    }
-
-    if (likely(!this_query)) {
-        ACLK_QUEUE_UNLOCK;
-        return NULL;
-    }
-
-    if (!this_query->deleted && this_query->run_after > now_realtime_sec()) {
-        info("Query %s will run in %ld seconds", this_query->query, this_query->run_after - now_realtime_sec());
-        ACLK_QUEUE_UNLOCK;
-        return NULL;
-    }
-
-    aclk_queue.count--;
-    aclk_queue.aclk_query_head = aclk_queue.aclk_query_head->next;
-
-    if (likely(!aclk_queue.aclk_query_head)) {
-        aclk_queue.aclk_query_tail = NULL;
-    }
-
-    ACLK_QUEUE_UNLOCK;
-    return this_query;
 }
 
 // This will give the base topic that the agent will publish messages.
@@ -768,277 +498,6 @@ void aclk_del_collector(const char *hostname, const char *plugin_name, const cha
 
     if (unlikely(aclk_queue_query("collector", NULL, NULL, NULL, 0, 1, ACLK_CMD_ONCONNECT)))
         debug(D_ACLK, "ACLK failed to queue on_connect command on collector deletion");
-}
-
-/*
- * Take a buffer, encode it and rewrite it
- *
- */
-
-static char *aclk_encode_response(char *src, size_t content_size, int keep_newlines)
-{
-    char *tmp_buffer = mallocz(content_size * 2);
-    char *dst = tmp_buffer;
-    while (content_size > 0) {
-        switch (*src) {
-            case '\n':
-                if (keep_newlines)
-                {
-                    *dst++ = '\\';
-                    *dst++ = 'n';
-                }
-                break;
-            case '\t':
-                break;
-            case 0x01 ... 0x08:
-            case 0x0b ... 0x1F:
-                *dst++ = '\\';
-                *dst++ = 'u';
-                *dst++ = '0';
-                *dst++ = '0';
-                *dst++ = (*src < 0x0F) ? '0' : '1';
-                *dst++ = to_hex(*src);
-                break;
-            case '\"':
-                *dst++ = '\\';
-                *dst++ = *src;
-                break;
-            default:
-                *dst++ = *src;
-        }
-        src++;
-        content_size--;
-    }
-    *dst = '\0';
-
-    return tmp_buffer;
-}
-
-int aclk_execute_query(struct aclk_query *this_query)
-{
-    if (strncmp(this_query->query, "/api/v1/", 8) == 0) {
-        struct web_client *w = (struct web_client *)callocz(1, sizeof(struct web_client));
-        w->response.data = buffer_create(NETDATA_WEB_RESPONSE_INITIAL_SIZE);
-        w->response.header = buffer_create(NETDATA_WEB_RESPONSE_HEADER_SIZE);
-        w->response.header_output = buffer_create(NETDATA_WEB_RESPONSE_HEADER_SIZE);
-        strcpy(w->origin, "*"); // Simulate web_client_create_on_fd()
-        w->cookie1[0] = 0;      // Simulate web_client_create_on_fd()
-        w->cookie2[0] = 0;      // Simulate web_client_create_on_fd()
-        w->acl = 0x1f;
-
-        char *mysep = strchr(this_query->query, '?');
-        if (mysep) {
-            strncpyz(w->decoded_query_string, mysep, NETDATA_WEB_REQUEST_URL_SIZE);
-            *mysep = '\0';
-        } else
-            strncpyz(w->decoded_query_string, this_query->query, NETDATA_WEB_REQUEST_URL_SIZE);
-
-        mysep = strrchr(this_query->query, '/');
-
-        // TODO: handle bad response perhaps in a different way. For now it does to the payload
-        w->response.code = web_client_api_request_v1(localhost, w, mysep ? mysep + 1 : "noop");
-        now_realtime_timeval(&w->tv_ready);
-        w->response.data->date = w->tv_ready.tv_sec;
-        web_client_build_http_header(w);  // TODO: this function should offset from date, not tv_ready
-        BUFFER *local_buffer = buffer_create(NETDATA_WEB_RESPONSE_INITIAL_SIZE);
-        buffer_flush(local_buffer);
-        local_buffer->contenttype = CT_APPLICATION_JSON;
-
-        aclk_create_header(local_buffer, "http", this_query->msg_id, 0, 0);
-        buffer_strcat(local_buffer, ",\n\t\"payload\": ");
-        char *encoded_response = aclk_encode_response(w->response.data->buffer, w->response.data->len, 0);
-        char *encoded_header = aclk_encode_response(w->response.header_output->buffer, w->response.header_output->len, 1);
-
-        buffer_sprintf(
-            local_buffer, "{\n\"code\": %d,\n\"body\": \"%s\",\n\"headers\": \"%s\"\n}", 
-            w->response.code, encoded_response, encoded_header);
-
-        buffer_sprintf(local_buffer, "\n}");
-
-        debug(D_ACLK, "Response:%s", encoded_header);
-
-        aclk_send_message(this_query->topic, local_buffer->buffer, this_query->msg_id);
-
-        buffer_free(w->response.data);
-        buffer_free(w->response.header);
-        buffer_free(w->response.header_output);
-        freez(w);
-        buffer_free(local_buffer);
-        freez(encoded_response);
-        freez(encoded_header);
-        return 0;
-    }
-    return 1;
-}
-
-/*
- * This function will fetch the next pending command and process it
- *
- */
-int aclk_process_query(int t_idx)
-{
-    struct aclk_query *this_query;
-    static long int query_count = 0;
-    ACLK_METADATA_STATE meta_state;
-
-    if (!aclk_connected)
-        return 0;
-
-    this_query = aclk_queue_pop();
-    if (likely(!this_query)) {
-        return 0;
-    }
-
-    if (unlikely(this_query->deleted)) {
-        debug(D_ACLK, "Garbage collect query %s:%s", this_query->topic, this_query->query);
-        aclk_query_free(this_query);
-        return 1;
-    }
-    query_count++;
-
-    debug(
-        D_ACLK, "Query #%ld (%s) size=%zu in queue %d seconds", query_count, this_query->topic,
-        this_query->query ? strlen(this_query->query) : 0, (int)(now_realtime_sec() - this_query->created));
-
-    switch (this_query->cmd) {
-        case ACLK_CMD_ONCONNECT:
-            debug(D_ACLK, "EXECUTING on connect metadata command");
-            ACLK_SHARED_STATE_LOCK;
-            meta_state = aclk_shared_state.metadata_submitted;
-            aclk_shared_state.metadata_submitted = ACLK_METADATA_SENT;
-            ACLK_SHARED_STATE_UNLOCK;
-            aclk_send_metadata(meta_state);
-            break;
-
-        case ACLK_CMD_CHART:
-            debug(D_ACLK, "EXECUTING a chart update command");
-            aclk_send_single_chart(this_query->data, this_query->query);
-            break;
-
-        case ACLK_CMD_CHARTDEL:
-            debug(D_ACLK, "EXECUTING a chart delete command");
-            //TODO: This send the info metadata for now
-            aclk_send_info_metadata();
-            break;
-
-        case ACLK_CMD_ALARM:
-            debug(D_ACLK, "EXECUTING an alarm update command");
-            aclk_send_message(this_query->topic, this_query->query, this_query->msg_id);
-            break;
-
-        case ACLK_CMD_CLOUD:
-            debug(D_ACLK, "EXECUTING a cloud command");
-            aclk_execute_query(this_query);
-            break;
-
-        default:
-            break;
-    }
-    debug(D_ACLK, "Query #%ld (%s) done", query_count, this_query->topic);
-
-    aclk_query_free(this_query);
-
-    if (aclk_stats_enabled) {
-        ACLK_STATS_LOCK;
-        aclk_metrics_per_sample.queries_dispatched++;
-        aclk_metrics_per_sample.queries_per_thread[t_idx]++;
-        ACLK_STATS_UNLOCK;
-    }
-
-    return 1;
-}
-
-/*
- * Process all pending queries
- * Return 0 if no queries were processed, 1 otherwise
- *
- */
-
-static inline int aclk_process_queries(int t_idx)
-{
-    if (unlikely(netdata_exit || !aclk_connected))
-        return 0;
-
-    if (likely(!aclk_queue.count))
-        return 0;
-
-    debug(D_ACLK, "Processing %d queries", (int)aclk_queue.count);
-
-    //TODO: may consider possible throttling here
-    while (aclk_process_query(t_idx)) {
-        // Process all commands
-    };
-
-    return 1;
-}
-
-/**
- * Main query processing thread
- *
- * On startup wait for the agent collectors to initialize
- * Expect at least a time of ACLK_STABLE_TIMEOUT seconds
- * of no new collectors coming in in order to mark the agent
- * as stable (set agent_state = AGENT_STABLE)
- */
-void *aclk_query_main_thread(void *ptr)
-{
-    struct aclk_query_thread *info = ptr;
-    time_t previous_popcorn_interrupt = 0;
-
-    while (!netdata_exit) {
-        ACLK_SHARED_STATE_LOCK;
-        if (aclk_shared_state.agent_state != AGENT_INITIALIZING) {
-            ACLK_SHARED_STATE_UNLOCK;
-            break;
-        }
-
-        time_t checkpoint = now_realtime_sec() - aclk_shared_state.last_popcorn_interrupt;
-
-        if (checkpoint > ACLK_STABLE_TIMEOUT) {
-            aclk_shared_state.agent_state = AGENT_STABLE;
-            ACLK_SHARED_STATE_UNLOCK;
-            info("AGENT stable, last collector initialization activity was %ld seconds ago", checkpoint);
-#ifdef ACLK_DEBUG
-            _dump_collector_list();
-#endif
-            break;
-        }
-
-        if (previous_popcorn_interrupt != aclk_shared_state.last_popcorn_interrupt) {
-            info("Waiting %ds from this moment for agent collectors to initialize." , ACLK_STABLE_TIMEOUT);
-            previous_popcorn_interrupt = aclk_shared_state.last_popcorn_interrupt;
-        }
-        ACLK_SHARED_STATE_UNLOCK;
-        sleep_usec(USEC_PER_SEC * 1);
-    }
-
-    while (!netdata_exit) {
-        ACLK_SHARED_STATE_LOCK;
-        if (unlikely(!aclk_shared_state.metadata_submitted)) {
-            ACLK_SHARED_STATE_UNLOCK;
-            if (unlikely(aclk_queue_query("on_connect", NULL, NULL, NULL, 0, 1, ACLK_CMD_ONCONNECT))) {
-                errno = 0;
-                error("ACLK failed to queue on_connect command");
-                sleep(1);
-                continue;
-            }
-            ACLK_SHARED_STATE_LOCK;
-            aclk_shared_state.metadata_submitted = ACLK_METADATA_CMD_QUEUED;
-        }
-        ACLK_SHARED_STATE_UNLOCK;
-
-        aclk_process_queries(info->idx);
-
-        QUERY_THREAD_LOCK;
-
-        // TODO: Need to check if there are queries awaiting already
-        if (unlikely(pthread_cond_wait(&query_cond_wait, &query_lock_wait)))
-            sleep_usec(USEC_PER_SEC * 1);
-
-        QUERY_THREAD_UNLOCK;
-    }
-
-    return NULL;
 }
 
 static void aclk_graceful_disconnect()
@@ -1385,53 +844,6 @@ static void aclk_try_to_connect(char *hostname, char *port, int port_num)
     }
 }
 
-static void aclk_query_threads_start(struct aclk_query_threads *query_threads)
-{
-    info("Starting %d query threads.", query_threads->count);
-
-#if ACLK_MAX_QUERY_THREADS > 10
-#error "You seem to have increased ACLK_MAX_QUERY_THREADS above 10. If you want to do that you need to update following thread name generation."
-#endif
-
-    char thread_name[strlen(ACLK_THREAD_NAME) + 3];
-    strcpy(thread_name, ACLK_THREAD_NAME);
-    thread_name[strlen(ACLK_THREAD_NAME)] = '_';
-    thread_name[strlen(ACLK_THREAD_NAME) + 2] = 0;
-
-    query_threads->thread_list = callocz(query_threads->count, sizeof(struct aclk_query_thread *));
-    for (int i = 0; i < query_threads->count; i++) {
-        query_threads->thread_list[i] = callocz(1, sizeof(struct aclk_query_thread));
-        query_threads->thread_list[i]->thread = callocz(1, sizeof(netdata_thread_t));
-        query_threads->thread_list[i]->idx = i; //thread needs to know its index for statistics
-
-        thread_name[strlen(ACLK_THREAD_NAME) + 1] = i + 0x30;
-        netdata_thread_create(
-            query_threads->thread_list[i]->thread, thread_name, NETDATA_THREAD_OPTION_JOINABLE, aclk_query_main_thread,
-            query_threads->thread_list[i]);
-    }
-}
-
-static void aclk_query_threads_cleanup(struct aclk_query_threads *query_threads)
-{
-    if (query_threads) {
-        for (int i = 0; i < query_threads->count; i++) {
-            netdata_thread_join(*(query_threads->thread_list[i]->thread), NULL);
-            freez(query_threads->thread_list[i]->thread);
-            freez(query_threads->thread_list[i]);
-        }
-    }
-
-    _reset_collector_list();
-    freez(collector_list);
-
-    struct aclk_query *this_query;
-
-    do {
-        this_query = aclk_queue_pop();
-        aclk_query_free(this_query);
-    } while (this_query);
-}
-
 /**
  * Main agent cloud link thread
  *
@@ -1604,6 +1016,9 @@ exited:
 
     aclk_query_threads_cleanup(&query_threads);
 
+    _reset_collector_list();
+    freez(collector_list);
+
     if(aclk_stats_enabled) {
         netdata_thread_join(*stats_thread->thread, NULL);
         freez(stats_thread->thread);
@@ -1701,7 +1116,8 @@ int aclk_subscribe(char *sub_topic, int qos)
 // This is called from a callback when the link goes up
 void aclk_connect()
 {
-    info("Connection detected (%"PRIu64" queued queries)", aclk_queue.count);
+    //TODO:underhood
+    //info("Connection detected (%"PRIu64" queued queries)", aclk_queue.count);
 
     aclk_stats_upd_online(1);
 
@@ -1715,8 +1131,9 @@ void aclk_connect()
 // This is called from a callback when the link goes down
 void aclk_disconnect()
 {
-    if (likely(aclk_connected))
-        info("Disconnect detected (%"PRIu64" queued queries)", aclk_queue.count);
+    //TODO:underhood
+    /*if (likely(aclk_connected))
+        info("Disconnect detected (%"PRIu64" queued queries)", aclk_queue.count);*/
 
     aclk_stats_upd_online(0);
